Remove Post list margin on left
In a previous change, we added `margin-left: 1.5em` on all `li,ul`
elements. Unfortunately, the post lists use `ul` so this master change
applied to every listing of posts too and now the website looks worse on
mobile.

Here we add special formatting for post lists to remove that global (and
generally sensible) margin