Improve Jekyll build time
I noticed after I added a bunch of source files to the example folder that
Jekyll was trying to *build* these files, and running a `make` in the
example directory was causing a very long rebuild.

After some exploration, it appears there's an `exclude` option we should be
using to exclude some of the folders.

https://github.com/jekyll/jekyll/blob/master/lib/site_template/_config.yml

I used to see build times times from around 20-30s, now I see 2-6 seconds.
Woo!