theme: fix pagination color text, enhance subscribe button
- When someone hovers the paginator, it doesn't highlight
the numbe with a different color. A ligther color has been
added not only for hover but for all other common used selectors.

- The subscribe button located at sidebar doesn't have a cursor
pointer when is hover. On top of that, when is clicked, an outline
appears. An  outline deletion and cursor pointer has been
added in order to fix these two behaviors.