Improved pagination UX (#191)
Hello 👋

I noticed that the blog page count has been increasing. Because of that, the pagination is starting to look overwhelming. Not that it looks bad at it's current state, but as the page count increases it will become overpopulated with page number buttons.

_Current_
![current pagination](https://i.imgur.com/SFxj8ci.png)

---

I have found a better solution for pagination. 

I took the inspiration from the following sources:
* https://dribbble.com/shots/2351417-What-s-your-fav-pagination (example number 7)
* https://uxframework.pearson.com/c/pagination (Standard pagination)
* https://codemyui.com/pagination-ux-pure-css/

Here's how it works:

The structure is as follows:
`[Previous page] [First page number] ... (N trailing) [Current page] (N leading) ... [Last page number] [Next page]`

There is a new option in **_config.yml** file called `paginate_trail` and the dafault value is `2`. It is used to generate N (in default case 2) trailing and leading page buttons.

_First page_
![first page](https://i.imgur.com/UposL6o.png)

You can notice that there is no "First page" button, because page 1 button is already visible.

_Second page_
![second page](https://i.imgur.com/4pD7AiU.png)

Second page also does not have the "First page" button, for same reason as for page 1.

_Page N_
![page n](https://i.imgur.com/ll6sS69.png)

When we get to page N, we can see trailing and leading N buttons (default 2), first page button (1), last page (10) and next and previous buttons. This way, user can proceed in reading posts chronologically, but he has an option to go to first and last page, as well as a way for "faster navigation" if a user is searing for some post.

_Last page_
![last page](https://i.imgur.com/jPlGBee.png)

The "Last page" button will not be visible on last page. Simillar logic for the first page.

---

I hope all of this makes sense. You can clone my fork and try it out locally.